BatterLevel now is more accurate
a mV level of 4600 should be around 60 and 5400 around 100

+++ 3pi_app11/3pi_app11/display.c
@@ -8,11 +8,11 @@
 #include "display.h"
 #include <pololu/3pi.h>
 
-int batteryPercentage()//Should this function be in main or in display.c or something else?
+float batteryPercentage()//Should this function be in main or in display.c or something else?
 {
 	
-	int milliVolts = read_battery_millivolts();
-	int batteryPercent = 0.000000000632*((milliVolts)*(milliVolts)*(milliVolts));
+	float milliVolts = read_battery_millivolts();
+	float batteryPercent = 0.000000000632*((milliVolts)*(milliVolts)*(milliVolts));
 	return batteryPercent;
 }
